Maybe fix publish

+++ publish.sh
@@ -4,13 +4,18 @@ set -e
 
 VERSION="$1"
 
+if [[ -z "$VERSION" ]]; then
+  echo "Need a version"
+  exit 1
+fi
+
 GOOS_LIST="linux darwin"
 GOARCH_LIST="amd64 arm64"
 
 rm -rf npm/bin/*
 for goos in $GOOS_LIST; do
   for goarch in $GOARCH_LIST; do
-    CGO_ENABLED=0 GOOS="$goos" GOARCH="$goarch" go build -ldflags="-s -w -X main.Version=$VERSION" -o "npm/bin/gopm3-${goos}-${goarch}"
+    CGO_ENABLED=0 GOOS="$goos" GOARCH="$goarch" go build -ldflags="-s -w -X main.Version=$VERSION" -o "npm/bin/gen-gopm3-${goos}-${goarch}"
   done
 done
 
@@ -19,12 +24,12 @@ cat << EOF > npm/package.json
   "name": "@hspak/gopm3",
   "version": "$VERSION",
   "description": "A dumb process manager",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "postinstall": "node install.js"
   },
   "bin": {
-    "gopm3": "./bin/gopm3"
+    "gopm3": "src/index.js"
   },
   "repository": {
     "type": "git",
@@ -41,3 +46,4 @@ EOF
 
 git tag "v${VERSION}" -f
 git push --follow-tags
+npm publish --access public npm
